Added some test base weapons to mon types
Monsters will have a base weapon as their default attack if they do not
have a weapon. Some monsters will only use this e.g. a ghoul will only
attack with claws. But for others, such as the player that will
typically be using weapons, the base "weapon" is just punching with
fists