Add fixed modifier to blocked path cost
Was returning infinity if a path is blocked by another creature. This
would cause a completely exhaustive search of all other paths if this
path is the only way to get to the destination.

Now just add a fixed 100 onto the cost, so that it will be evaluated
faster before exhausting everything else