Added orange colours and mob DC colours
To represent the threat level of symbols, there are 7 coloured DC levels.

These are: grey < green < blue < yellow < orange < red < purple

Grey enemies are the weakest types and purple are the strongest. Though
this is just an indication, it should always be possible for even a high
level player to be killed by grey enemies if they are not careful.
Purple enemies represent very tough enemies (i.e. bosses) and should
always be approached with caution.