Updated mapgen to take a map struct to generate
Mapgen was only working on cmap (current map)
This update allows other maps to be generated even if they are not the
current map, for future generation of adjacent maps.

Also updated internal mapgen functions to return the location where they were
originally assigning to an out pointer and returning a bool.