Connected cells together
Created a method to generate a map entirely rather than cell by cell.
This also allows a "connect cells" function to be called easily.

Created a "connect cells" function that just trivially carves a random
amount of paths orthogonally between cells.

Now that "seamless" map cells are a thing, the map generation takes lot
longer, so added a little bit of text to tell the player that the map is
being generated.

Set player health to 999 for easier debugging