Added initial commit of tasking
The first bit of multithreading in naxx history! The tasker is designed
to sit there on a separate thread processing "tasks", it will sleep if
there are no tasks to process.

The tasking system is cooperative, and task execution functions must be
designed with that philosophy in mind. A task is a wrapper around an
execution function, its arguments, and a callback function. The task
args are memcpy'd on task creation so that must also be kept in mind.